added icons for Classs and Archetype features
Notes on icon symbols:
Square outline = feature that is an Action
Diagonaly cut in half Half Square outline = feature that is a Bonus Action
Corner of a square = feature is a reaction.
Small square in bottom right corner = feature is passive or otherwise special

Class images are shown for each one.
Colors represent Base class or archetype as follows:
All base classes: Grey
Archetype monocolor for:
	Berserker - red
	Consular - green
	Engineer - brown
	Fighter - orange
	Monk - white
	Operative - black
	Scholar - violet
	Scout - spring green (i dub it Spriggeen)
	Guardian - azure
	Sentinel - yellow
Finally the symbols in the bottom left of each icon are Aurebesh letters, and are simply the first letter of either Action, Bonus, Reaction or Passive respectivly.