Starship Actions Update + Clear Midi Flags
Updated Boosts to be DC10 instead of DC12

Deleted Cast a Power as not an action on website

Deleted the last line in the Conceal description that wasn’t on the website

Added Detect

Deleted the last line in the Dogfight description that wasn’t on the website

Deleted the last line in the Evade description that wasn’t on the website

Updated last sentence in Fire to match website

Updated 2nd paragraph in Fly and deleted last line that wasn’t on the website

Updated Harass description to match website

Updated Interfere description to match website

Added Hard Turn

Updated Patch description to match website

Deleted the last line in the Ram description that wasn’t on the website

Deleted Ready as not an action on website

Added last sentence from website to Regenerate Shields. Also added 1 Turn duration to match the sentence terms.

Deleted Search since Detect took its place

Deleted Use an Object as not an action on website